pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "arches-lingo"
readme = "README.md"
authors = [
    {"name" = "Arches Project"}
]
license = "AGPL-3.0-or-later"
license-files = ["LICENSE"]
description = "An Arches application designed to support the creation of thesaurus/vocabulary data."
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
]
requires-python = ">=3.11"
dependencies = [
    "arches @ git+https://github.com/archesproject/arches@dev/8.1.x",
    "arches_component_lab @ git+https://github.com/archesproject/arches-component-lab@main",
    "arches_querysets @ git+https://github.com/archesproject/arches-querysets@dev/1.1.x",
    "arches_controlled_lists @ git+https://github.com/archesproject/arches-controlled-lists@dev/1.1.x",
]
version = "1.0.0a0"

[project.urls]
Homepage = "https://archesproject.org/"
Documentation = "https://arches.readthedocs.io"
Repository = "https://github.com/archesproject/arches-lingo.git"
Issues = "https://github.com/archesproject/arches-lingo/issues"

[dependency-groups]
dev = [
    "livereload",
    "sst",
    "coverage",
    "django-silk==5.1.0",
    "pre-commit",
    "black==24.4.2",
]

[tool.setuptools]
packages = ["arches_lingo"]
